[kyo-schema] extract kyo-schema-protobuf
