+++ kyo-schema-json/shared/src/test/scala/kyo/SchemaNamingTest.scala
