+++ kyo-schema-protobuf/shared/src/main/scala/kyo/internal/ProtobufReader.scala
