+++ kyo-schema-tests/shared/src/test/scala/kyo/BuilderTest.scala
