+++ kyo-schema-tests/shared/src/test/scala/kyo/SchemaUnionRepresentationTest.scala
